Add \sized macro for symbol nodes
This allows you to create symbol nodes that whose width is determined by
some other symbol than the one shown in the node itself.  That lets you
create several nodes that all have the same size.  There are two ways to
use it:

    % makes the node big enough to hold "large_symbol", but only shows
    % "symbol"
    \sized[large\_symbol]{symbol}

If there are going to be several nodes in a row that all use the same
“sizing symbol”, you can use the \setwidestsymbol command to only have
ot specify it once:

    % same effect as above
    \setwidestsymbol{large\_symbol}
    \sized{symbol}